Refactor range and keys getting
