+++ src/utils/table.ts
@@ -13,6 +13,9 @@
 // limitations under the License.
 
 import {GetRowsOptions, PrefixRange, Table} from '../table';
+import {Mutation} from '../mutation';
+import * as is from 'is';
+import {BoundData} from '../filter';
 
 export class TableUtils {
   static getRanges(options: GetRowsOptions) {
@@ -47,6 +50,58 @@ export class TableUtils {
       });
     }
     return ranges;
+  }
+
+  static spliceRangesGetKeys(
+    ranges: PrefixRange[],
+    lastRowKey: string | number | true | Uint8Array,
+    rowKeys: string[]
+  ): string[] {
+    // TODO: lhs and rhs type shouldn't be string, it could be
+    // string, number, Uint8Array, boolean. Fix the type
+    // and clean up the casting.
+    const lessThan = (lhs: string, rhs: string) => {
+      const lhsBytes = Mutation.convertToBytes(lhs);
+      const rhsBytes = Mutation.convertToBytes(rhs);
+      return (lhsBytes as Buffer).compare(rhsBytes as Uint8Array) === -1;
+    };
+    const greaterThan = (lhs: string, rhs: string) => lessThan(rhs, lhs);
+    const lessThanOrEqualTo = (lhs: string, rhs: string) =>
+      !greaterThan(lhs, rhs);
+
+    // Readjust and/or remove ranges based on previous valid row reads.
+    // Iterate backward since items may need to be removed.
+    for (let index = ranges.length - 1; index >= 0; index--) {
+      const range = ranges[index];
+      const startValue = is.object(range.start)
+        ? (range.start as BoundData).value
+        : range.start;
+      const endValue = is.object(range.end)
+        ? (range.end as BoundData).value
+        : range.end;
+      const startKeyIsRead =
+        !startValue ||
+        lessThanOrEqualTo(startValue as string, lastRowKey as string);
+      const endKeyIsNotRead =
+        !endValue ||
+        (endValue as Buffer).length === 0 ||
+        lessThan(lastRowKey as string, endValue as string);
+      if (startKeyIsRead) {
+        if (endKeyIsNotRead) {
+          // EndKey is not read, reset the range to start from lastRowKey open
+          range.start = {
+            value: lastRowKey,
+            inclusive: false,
+          };
+        } else {
+          // EndKey is read, remove this range
+          ranges.splice(index, 1);
+        }
+      }
+    }
+
+    // Remove rowKeys already read.
+    return rowKeys.filter(rowKey => greaterThan(rowKey, lastRowKey as string));
   }
 
   static createPrefixRange(start: string): PrefixRange {
